Run performance supports existing org with dynamic spaces
