+++ src/acceptance/setup_performance/setup_performance_suite_test.go
@@ -37,7 +37,7 @@ var _ = BeforeSuite(func() {
 	if os.Getenv("SKIP_TEARDOWN") == "true" {
 		fmt.Println("Skipping Teardown...")
 	} else {
-		cleanupExistingSpaceIfEnabled()
+		cleanup()
 	}
 
 	setup = workflowhelpers.NewRunawayAppTestSuiteSetup(cfg)
@@ -92,25 +92,14 @@ func deleteExistingServiceInstances(workerId int, servicesChan chan string, setu
 	fmt.Printf("worker %d  - Delete Service Instance finished...\n", workerId)
 }
 
-func cleanupExistingSpaceIfEnabled() {
-	if !cfg.UseExistingSpace {
-		fmt.Printf("\nusing existing space = %t ..skipping space deletion", cfg.UseExistingSpace)
-		return
-	}
+func cleanup() {
 	setup = workflowhelpers.NewTestSuiteSetup(cfg)
-
 	workflowhelpers.AsUser(setup.AdminUserContext(), cfg.DefaultTimeoutDuration(), func() {
-		organizations := GetTestOrgs(cfg)
-		Expect(len(organizations)).To(Equal(1))
-		orgName := organizations[0]
-		orgGuid := GetOrgGuid(cfg, orgName)
+		DeleteOrgs(GetTestOrgs(cfg), time.Duration(120)*time.Second)
 
 		if cfg.UseExistingOrganization {
-			spaces := GetTestSpaces(orgGuid, cfg)
-			DeleteSpaces(orgName, spaces, cfg.DefaultTimeoutDuration())
-		} else {
-			orgName := setup.GetOrganizationName()
-			DeleteOrgWithTimeout(orgName, time.Duration(120)*time.Second)
+			orgGuid := GetOrgGuid(cfg, cfg.ExistingOrganization)
+			DeleteSpaces(cfg.ExistingOrganization, GetTestSpaces(orgGuid, cfg), cfg.DefaultTimeoutDuration())
 		}
 	})
